separated classes into different files in the same namespace beginnerApp
